+++ packages/orq-rc/src/orq_ai_sdk/models/__init__.py
@@ -432,6 +432,8 @@ if TYPE_CHECKING:
         InputRole,
         InputStatus,
         InputType,
+        MCPTool,
+        MCPToolTypedDict,
         Network,
         NetworkTypedDict,
         OrqAiTool,
@@ -448,8 +450,6 @@ if TYPE_CHECKING:
         ToolsFilesTypedDict,
         ToolsFunction,
         ToolsFunctionTypedDict,
-        ToolsMCPTool,
-        ToolsMCPToolTypedDict,
         ToolsMode,
         ToolsTTL,
         ToolsType,
@@ -1765,30 +1765,6 @@ if TYPE_CHECKING:
         CreateManagementKeyResponse,
         CreateManagementKeyResponseTypedDict,
     )
-    from .createmcpgatewayrequest import (
-        CreateMcpGatewayRequest,
-        CreateMcpGatewayRequestTypedDict,
-    )
-    from .createmcpgatewayresponse import (
-        CreateMcpGatewayResponse,
-        CreateMcpGatewayResponseTypedDict,
-    )
-    from .createmcpserverrequest import (
-        CreateMcpServerRequest,
-        CreateMcpServerRequestTypedDict,
-    )
-    from .createmcpserverresponse import (
-        CreateMcpServerResponse,
-        CreateMcpServerResponseTypedDict,
-    )
-    from .createmcptoolsetrequest import (
-        CreateMcpToolsetRequest,
-        CreateMcpToolsetRequestTypedDict,
-    )
-    from .createmcptoolsetresponse import (
-        CreateMcpToolsetResponse,
-        CreateMcpToolsetResponseTypedDict,
-    )
     from .creatememorydocumentop import (
         CreateMemoryDocumentRequest,
         CreateMemoryDocumentRequestBody,
@@ -2585,18 +2561,6 @@ if TYPE_CHECKING:
     from .deletemanagementkeyresponse import (
         DeleteManagementKeyResponse,
         DeleteManagementKeyResponseTypedDict,
-    )
-    from .deletemcpgatewayresponse import (
-        DeleteMcpGatewayResponse,
-        DeleteMcpGatewayResponseTypedDict,
-    )
-    from .deletemcpserverresponse import (
-        DeleteMcpServerResponse,
-        DeleteMcpServerResponseTypedDict,
-    )
-    from .deletemcptoolsetresponse import (
-        DeleteMcpToolsetResponse,
-        DeleteMcpToolsetResponseTypedDict,
     )
     from .deletememorydocumentop import (
         DeleteMemoryDocumentRequest,
@@ -3929,18 +3893,6 @@ if TYPE_CHECKING:
         GetManagementKeyResponse,
         GetManagementKeyResponseTypedDict,
     )
-    from .getmcpgatewayresponse import (
-        GetMcpGatewayResponse,
-        GetMcpGatewayResponseTypedDict,
-    )
-    from .getmcpserverresponse import (
-        GetMcpServerResponse,
-        GetMcpServerResponseTypedDict,
-    )
-    from .getmcptoolsetresponse import (
-        GetMcpToolsetResponse,
-        GetMcpToolsetResponseTypedDict,
-    )
     from .getmonitorresponse import GetMonitorResponse, GetMonitorResponseTypedDict
     from .getnotifierresponse import GetNotifierResponse, GetNotifierResponseTypedDict
     from .getonechunkop import (
@@ -5159,30 +5111,6 @@ if TYPE_CHECKING:
         ListManagementKeysResponse,
         ListManagementKeysResponseTypedDict,
     )
-    from .listmcpgatewaycallsresponse import (
-        ListMcpGatewayCallsResponse,
-        ListMcpGatewayCallsResponseTypedDict,
-    )
-    from .listmcpgatewaysresponse import (
-        ListMcpGatewaysResponse,
-        ListMcpGatewaysResponseTypedDict,
-    )
-    from .listmcpgatewaytoolsresponse import (
-        ListMcpGatewayToolsResponse,
-        ListMcpGatewayToolsResponseTypedDict,
-    )
-    from .listmcpserversresponse import (
-        ListMcpServersResponse,
-        ListMcpServersResponseTypedDict,
-    )
-    from .listmcpsessionsresponse import (
-        ListMcpSessionsResponse,
-        ListMcpSessionsResponseTypedDict,
-    )
-    from .listmcptoolsetsresponse import (
-        ListMcpToolsetsResponse,
-        ListMcpToolsetsResponseTypedDict,
-    )
     from .listmonitorpresetsresponse import (
         ListMonitorPresetsResponse,
         ListMonitorPresetsResponseTypedDict,
@@ -5434,103 +5362,7 @@ if TYPE_CHECKING:
     )
     from .managementpermissionmode import ManagementPermissionMode
     from .mcpaccess import McpAccess, McpAccessTypedDict
-    from .mcpauthconfig import McpAuthConfig, McpAuthConfigTypedDict
-    from .mcpauthtype import McpAuthType
-    from .mcpconnection import McpConnection, McpConnectionTypedDict
-    from .mcpconnectiontype import McpConnectionType
-    from .mcpegresspolicy import McpEgressPolicy, McpEgressPolicyTypedDict
-    from .mcpgateway import McpGateway, McpGatewayTypedDict
-    from .mcpgatewaycall import McpGatewayCall, McpGatewayCallTypedDict
-    from .mcpgatewaycalllistop import (
-        McpGatewayCallListRequest,
-        McpGatewayCallListRequestTypedDict,
-    )
-    from .mcpgatewaycallstatus import McpGatewayCallStatus
-    from .mcpgatewaydeleteop import (
-        McpGatewayDeleteRequest,
-        McpGatewayDeleteRequestTypedDict,
-    )
-    from .mcpgatewaygetop import McpGatewayGetRequest, McpGatewayGetRequestTypedDict
-    from .mcpgatewaylistop import McpGatewayListRequest, McpGatewayListRequestTypedDict
-    from .mcpgatewaylisttoolsop import (
-        McpGatewayListToolsRequest,
-        McpGatewayListToolsRequestTypedDict,
-    )
-    from .mcpgatewaymode import McpGatewayMode
-    from .mcpgatewayserverlink import (
-        McpGatewayServerLink,
-        McpGatewayServerLinkTypedDict,
-    )
-    from .mcpgatewaystatus import McpGatewayStatus
-    from .mcpgatewaytool import (
-        InputSchema,
-        InputSchemaTypedDict,
-        McpGatewayTool,
-        McpGatewayToolTypedDict,
-    )
-    from .mcpgatewaytooltestresult import (
-        McpGatewayToolTestResult,
-        McpGatewayToolTestResultTypedDict,
-        Result,
-        ResultTypedDict,
-    )
-    from .mcpgatewayupdateop import (
-        McpGatewayUpdateRequest,
-        McpGatewayUpdateRequestTypedDict,
-    )
-    from .mcpheadersecret import McpHeaderSecret, McpHeaderSecretTypedDict
-    from .mcpoauthconfig import McpOAuthConfig, McpOAuthConfigTypedDict
-    from .mcpruntimelimits import McpRuntimeLimits, McpRuntimeLimitsTypedDict
-    from .mcpserver import McpServer, McpServerTypedDict
-    from .mcpserverdeleteop import (
-        McpServerDeleteRequest,
-        McpServerDeleteRequestTypedDict,
-    )
-    from .mcpservergetop import McpServerGetRequest, McpServerGetRequestTypedDict
-    from .mcpserverlistop import McpServerListRequest, McpServerListRequestTypedDict
-    from .mcpserversharing import McpServerSharing, McpServerSharingTypedDict
-    from .mcpserversharingscope import McpServerSharingScope
-    from .mcpserverstatus import McpServerStatus
-    from .mcpserversyncop import McpServerSyncRequest, McpServerSyncRequestTypedDict
-    from .mcpservertesttoolop import (
-        McpServerTestToolRequest,
-        McpServerTestToolRequestTypedDict,
-    )
-    from .mcpserverupdateop import (
-        McpServerUpdateRequest,
-        McpServerUpdateRequestTypedDict,
-    )
-    from .mcpsession import McpSession, McpSessionTypedDict
-    from .mcpsessionlistop import McpSessionListRequest, McpSessionListRequestTypedDict
-    from .mcpsessionrevokeop import (
-        McpSessionRevokeRequest,
-        McpSessionRevokeRequestTypedDict,
-    )
-    from .mcpsessionstatus import McpSessionStatus
-    from .mcpsyncstate import McpSyncState, McpSyncStateTypedDict
-    from .mcptestresult import McpTestResult, McpTestResultTypedDict
-    from .mcptool import (
-        McpTool,
-        McpToolInputSchema,
-        McpToolInputSchemaTypedDict,
-        McpToolTypedDict,
-    )
-    from .mcptoolexposure import McpToolExposure, McpToolExposureTypedDict
-    from .mcptoolexposuremode import McpToolExposureMode
     from .mcptoolinput import McpToolInput, McpToolInputType, McpToolInputTypedDict
-    from .mcptoolnaming import McpToolNaming
-    from .mcptoolset import McpToolset, McpToolsetTypedDict
-    from .mcptoolsetdeleteop import (
-        McpToolsetDeleteRequest,
-        McpToolsetDeleteRequestTypedDict,
-    )
-    from .mcptoolsetgetop import McpToolsetGetRequest, McpToolsetGetRequestTypedDict
-    from .mcptoolsetlistop import McpToolsetListRequest, McpToolsetListRequestTypedDict
-    from .mcptoolsettoolref import McpToolsetToolRef, McpToolsetToolRefTypedDict
-    from .mcptoolsetupdateop import (
-        McpToolsetUpdateRequest,
-        McpToolsetUpdateRequestTypedDict,
-    )
     from .memoryparam import MemoryParam, MemoryParamTypedDict
     from .modelazurefoundrydeploymentsop import (
         ModelAzureFoundryDeploymentsRequestBody,
@@ -6750,10 +6582,6 @@ if TYPE_CHECKING:
         RetrieveToolResponseBodyTypedDict,
     )
     from .reviewoutcome import ReviewOutcome
-    from .revokemcpsessionresponse import (
-        RevokeMcpSessionResponse,
-        RevokeMcpSessionResponseTypedDict,
-    )
     from .routingrule import RoutingRule, RoutingRuleTypedDict
     from .routingrulecreateop import (
         RoutingRuleCreateRequestBody,
@@ -6833,6 +6661,8 @@ if TYPE_CHECKING:
         AgentToolInputRunWebScraperToolTypedDict,
         AgentToolInputRunWriteMemoryStoreTool,
         AgentToolInputRunWriteMemoryStoreToolTypedDict,
+        Arguments,
+        ArgumentsTypedDict,
         Blueprint,
         BlueprintTypedDict,
         CodeTool,
@@ -6881,8 +6711,6 @@ if TYPE_CHECKING:
         RunAgentAgentToolInputRunAgentsRequestRequestBodyType,
         RunAgentAgentToolInputRunAgentsRequestType,
         RunAgentAgentToolInputRunAgentsType,
-        RunAgentAgentToolInputRunArguments,
-        RunAgentAgentToolInputRunArgumentsTypedDict,
         RunAgentAgentToolInputRunHeaders,
         RunAgentAgentToolInputRunHeadersTypedDict,
         RunAgentAgentToolInputRunParameters,
@@ -7506,35 +7334,8 @@ if TYPE_CHECKING:
         StreamRunAgentToolChoiceFunctionTypedDict,
         StreamRunAgentToolChoiceType,
     )
-    from .syncmcpserverrequest import (
-        SyncMcpServerRequest,
-        SyncMcpServerRequestTypedDict,
-    )
-    from .syncmcpserverresponse import (
-        SyncMcpServerResponse,
-        SyncMcpServerResponseTypedDict,
-    )
-    from .syncstatus import SyncStatus
     from .telemetry import Telemetry, TelemetryTypedDict
     from .telemetryrow import TelemetryRow, TelemetryRowTypedDict
-    from .testmcpserverrequest import (
-        TestMcpServerRequest,
-        TestMcpServerRequestTypedDict,
-    )
-    from .testmcpserverresponse import (
-        TestMcpServerResponse,
-        TestMcpServerResponseTypedDict,
-    )
-    from .testmcpservertoolrequest import (
-        Arguments,
-        ArgumentsTypedDict,
-        TestMcpServerToolRequest,
-        TestMcpServerToolRequestTypedDict,
-    )
-    from .testmcpservertoolresponse import (
-        TestMcpServerToolResponse,
-        TestMcpServerToolResponseTypedDict,
-    )
     from .textcontentpartschema import (
         CacheControl,
         CacheControlTypedDict,
@@ -8290,30 +8091,6 @@ if TYPE_CHECKING:
     from .updatemanagementkeyresponse import (
         UpdateManagementKeyResponse,
         UpdateManagementKeyResponseTypedDict,
-    )
-    from .updatemcpgatewayrequest import (
-        UpdateMcpGatewayRequest,
-        UpdateMcpGatewayRequestTypedDict,
-    )
-    from .updatemcpgatewayresponse import (
-        UpdateMcpGatewayResponse,
-        UpdateMcpGatewayResponseTypedDict,
-    )
-    from .updatemcpserverrequest import (
-        UpdateMcpServerRequest,
-        UpdateMcpServerRequestTypedDict,
-    )
-    from .updatemcpserverresponse import (
-        UpdateMcpServerResponse,
-        UpdateMcpServerResponseTypedDict,
-    )
-    from .updatemcptoolsetrequest import (
-        UpdateMcpToolsetRequest,
-        UpdateMcpToolsetRequestTypedDict,
-    )
-    from .updatemcptoolsetresponse import (
-        UpdateMcpToolsetResponse,
-        UpdateMcpToolsetResponseTypedDict,
     )
     from .updatememorydocumentop import (
         UpdateMemoryDocumentRequest,
@@ -10326,18 +10103,6 @@ __all__ = [
     "CreateManagementKeyRequestTypedDict",
     "CreateManagementKeyResponse",
     "CreateManagementKeyResponseTypedDict",
-    "CreateMcpGatewayRequest",
-    "CreateMcpGatewayRequestTypedDict",
-    "CreateMcpGatewayResponse",
-    "CreateMcpGatewayResponseTypedDict",
-    "CreateMcpServerRequest",
-    "CreateMcpServerRequestTypedDict",
-    "CreateMcpServerResponse",
-    "CreateMcpServerResponseTypedDict",
-    "CreateMcpToolsetRequest",
-    "CreateMcpToolsetRequestTypedDict",
-    "CreateMcpToolsetResponse",
-    "CreateMcpToolsetResponseTypedDict",
     "CreateMemoryDocumentRequest",
     "CreateMemoryDocumentRequestBody",
     "CreateMemoryDocumentRequestBodyTypedDict",
@@ -11070,12 +10835,6 @@ __all__ = [
     "DeleteKnowledgeRequestTypedDict",
     "DeleteManagementKeyResponse",
     "DeleteManagementKeyResponseTypedDict",
-    "DeleteMcpGatewayResponse",
-    "DeleteMcpGatewayResponseTypedDict",
-    "DeleteMcpServerResponse",
-    "DeleteMcpServerResponseTypedDict",
-    "DeleteMcpToolsetResponse",
-    "DeleteMcpToolsetResponseTypedDict",
     "DeleteMemoryDocumentRequest",
     "DeleteMemoryDocumentRequestTypedDict",
     "DeleteMemoryDocumentTool",
@@ -12263,12 +12022,6 @@ __all__ = [
     "GetLogResponseTypedDict",
     "GetManagementKeyResponse",
     "GetManagementKeyResponseTypedDict",
-    "GetMcpGatewayResponse",
-    "GetMcpGatewayResponseTypedDict",
-    "GetMcpServerResponse",
-    "GetMcpServerResponseTypedDict",
-    "GetMcpToolsetResponse",
-    "GetMcpToolsetResponseTypedDict",
     "GetMonitorResponse",
     "GetMonitorResponseTypedDict",
     "GetNotifierResponse",
@@ -12800,8 +12553,6 @@ __all__ = [
     "InputMessage",
     "InputMessageTypedDict",
     "InputRole",
-    "InputSchema",
-    "InputSchemaTypedDict",
     "InputStatus",
     "InputTokensDetails",
     "InputTokensDetailsTypedDict",
@@ -13331,18 +13082,6 @@ __all__ = [
     "ListManagementCapabilitiesResponseTypedDict",
     "ListManagementKeysResponse",
     "ListManagementKeysResponseTypedDict",
-    "ListMcpGatewayCallsResponse",
-    "ListMcpGatewayCallsResponseTypedDict",
-    "ListMcpGatewayToolsResponse",
-    "ListMcpGatewayToolsResponseTypedDict",
-    "ListMcpGatewaysResponse",
-    "ListMcpGatewaysResponseTypedDict",
-    "ListMcpServersResponse",
-    "ListMcpServersResponseTypedDict",
-    "ListMcpSessionsResponse",
-    "ListMcpSessionsResponseTypedDict",
-    "ListMcpToolsetsResponse",
-    "ListMcpToolsetsResponseTypedDict",
     "ListModelsObject",
     "ListModelsResponseBody",
     "ListModelsResponseBodyTypedDict",
@@ -13576,8 +13315,10 @@ __all__ = [
     "LtTypedDict",
     "Lte",
     "LteTypedDict",
+    "MCPTool",
     "MCPToolRun",
     "MCPToolRunTypedDict",
+    "MCPToolTypedDict",
     "ManagementKey",
     "ManagementKeyDeleteRequest",
     "ManagementKeyDeleteRequestTypedDict",
@@ -13595,97 +13336,9 @@ __all__ = [
     "Mcp",
     "McpAccess",
     "McpAccessTypedDict",
-    "McpAuthConfig",
-    "McpAuthConfigTypedDict",
-    "McpAuthType",
-    "McpConnection",
-    "McpConnectionType",
-    "McpConnectionTypedDict",
-    "McpEgressPolicy",
-    "McpEgressPolicyTypedDict",
-    "McpGateway",
-    "McpGatewayCall",
-    "McpGatewayCallListRequest",
-    "McpGatewayCallListRequestTypedDict",
-    "McpGatewayCallStatus",
-    "McpGatewayCallTypedDict",
-    "McpGatewayDeleteRequest",
-    "McpGatewayDeleteRequestTypedDict",
-    "McpGatewayGetRequest",
-    "McpGatewayGetRequestTypedDict",
-    "McpGatewayListRequest",
-    "McpGatewayListRequestTypedDict",
-    "McpGatewayListToolsRequest",
-    "McpGatewayListToolsRequestTypedDict",
-    "McpGatewayMode",
-    "McpGatewayServerLink",
-    "McpGatewayServerLinkTypedDict",
-    "McpGatewayStatus",
-    "McpGatewayTool",
-    "McpGatewayToolTestResult",
-    "McpGatewayToolTestResultTypedDict",
-    "McpGatewayToolTypedDict",
-    "McpGatewayTypedDict",
-    "McpGatewayUpdateRequest",
-    "McpGatewayUpdateRequestTypedDict",
-    "McpHeaderSecret",
-    "McpHeaderSecretTypedDict",
-    "McpOAuthConfig",
-    "McpOAuthConfigTypedDict",
-    "McpRuntimeLimits",
-    "McpRuntimeLimitsTypedDict",
-    "McpServer",
-    "McpServerDeleteRequest",
-    "McpServerDeleteRequestTypedDict",
-    "McpServerGetRequest",
-    "McpServerGetRequestTypedDict",
-    "McpServerListRequest",
-    "McpServerListRequestTypedDict",
-    "McpServerSharing",
-    "McpServerSharingScope",
-    "McpServerSharingTypedDict",
-    "McpServerStatus",
-    "McpServerSyncRequest",
-    "McpServerSyncRequestTypedDict",
-    "McpServerTestToolRequest",
-    "McpServerTestToolRequestTypedDict",
-    "McpServerTypedDict",
-    "McpServerUpdateRequest",
-    "McpServerUpdateRequestTypedDict",
-    "McpSession",
-    "McpSessionListRequest",
-    "McpSessionListRequestTypedDict",
-    "McpSessionRevokeRequest",
-    "McpSessionRevokeRequestTypedDict",
-    "McpSessionStatus",
-    "McpSessionTypedDict",
-    "McpSyncState",
-    "McpSyncStateTypedDict",
-    "McpTestResult",
-    "McpTestResultTypedDict",
-    "McpTool",
-    "McpToolExposure",
-    "McpToolExposureMode",
-    "McpToolExposureTypedDict",
     "McpToolInput",
-    "McpToolInputSchema",
-    "McpToolInputSchemaTypedDict",
     "McpToolInputType",
     "McpToolInputTypedDict",
-    "McpToolNaming",
-    "McpToolTypedDict",
-    "McpToolset",
-    "McpToolsetDeleteRequest",
-    "McpToolsetDeleteRequestTypedDict",
-    "McpToolsetGetRequest",
-    "McpToolsetGetRequestTypedDict",
-    "McpToolsetListRequest",
-    "McpToolsetListRequestTypedDict",
-    "McpToolsetToolRef",
-    "McpToolsetToolRefTypedDict",
-    "McpToolsetTypedDict",
-    "McpToolsetUpdateRequest",
-    "McpToolsetUpdateRequestTypedDict",
     "McpTypedDict",
     "Memory",
     "MemoryParam",
@@ -14682,8 +14335,6 @@ __all__ = [
     "RestoreRequestTypedDict",
     "RestoreResponse",
     "RestoreResponseTypedDict",
-    "Result",
-    "ResultTypedDict",
     "Results",
     "Results1",
     "Results1TypedDict",
@@ -15060,8 +14711,6 @@ __all__ = [
     "ReturnType",
     "Review",
     "ReviewOutcome",
-    "RevokeMcpSessionResponse",
-    "RevokeMcpSessionResponseTypedDict",
     "Role",
     "RoleToolMessage",
     "RoleUserMessage",
@@ -15122,8 +14771,6 @@ __all__ = [
     "RunAgentAgentToolInputRunAgentsRequestRequestBodyType",
     "RunAgentAgentToolInputRunAgentsRequestType",
     "RunAgentAgentToolInputRunAgentsType",
-    "RunAgentAgentToolInputRunArguments",
-    "RunAgentAgentToolInputRunArgumentsTypedDict",
     "RunAgentAgentToolInputRunHeaders",
     "RunAgentAgentToolInputRunHeadersTypedDict",
     "RunAgentAgentToolInputRunParameters",
@@ -15747,11 +15394,6 @@ __all__ = [
     "StructuredTypedDict",
     "Style",
     "Summary",
-    "SyncMcpServerRequest",
-    "SyncMcpServerRequestTypedDict",
-    "SyncMcpServerResponse",
-    "SyncMcpServerResponseTypedDict",
-    "SyncStatus",
     "SystemMessage",
     "SystemMessageTypedDict",
     "TTL",
@@ -15769,14 +15411,6 @@ __all__ = [
     "TemplateEngine",
     "Ten",
     "TenTypedDict",
-    "TestMcpServerRequest",
-    "TestMcpServerRequestTypedDict",
-    "TestMcpServerResponse",
-    "TestMcpServerResponseTypedDict",
-    "TestMcpServerToolRequest",
-    "TestMcpServerToolRequestTypedDict",
-    "TestMcpServerToolResponse",
-    "TestMcpServerToolResponseTypedDict",
     "Text",
     "TextContentPartSchema",
     "TextContentPartSchemaCacheControlType",
@@ -15909,8 +15543,6 @@ __all__ = [
     "ToolsFilesTypedDict",
     "ToolsFunction",
     "ToolsFunctionTypedDict",
-    "ToolsMCPTool",
-    "ToolsMCPToolTypedDict",
     "ToolsMode",
     "ToolsModel",
     "ToolsModelTypedDict",
@@ -16565,18 +16197,6 @@ __all__ = [
     "UpdateManagementKeyRequestTypedDict",
     "UpdateManagementKeyResponse",
     "UpdateManagementKeyResponseTypedDict",
-    "UpdateMcpGatewayRequest",
-    "UpdateMcpGatewayRequestTypedDict",
-    "UpdateMcpGatewayResponse",
-    "UpdateMcpGatewayResponseTypedDict",
-    "UpdateMcpServerRequest",
-    "UpdateMcpServerRequestTypedDict",
-    "UpdateMcpServerResponse",
-    "UpdateMcpServerResponseTypedDict",
-    "UpdateMcpToolsetRequest",
-    "UpdateMcpToolsetRequestTypedDict",
-    "UpdateMcpToolsetResponse",
-    "UpdateMcpToolsetResponseTypedDict",
     "UpdateMemoryDocumentRequest",
     "UpdateMemoryDocumentRequestBody",
     "UpdateMemoryDocumentRequestBodyTypedDict",
@@ -17501,6 +17121,8 @@ _dynamic_imports: dict[str, str] = {
     "InputRole": ".create_router_responseop",
     "InputStatus": ".create_router_responseop",
     "InputType": ".create_router_responseop",
+    "MCPTool": ".create_router_responseop",
+    "MCPToolTypedDict": ".create_router_responseop",
     "Network": ".create_router_responseop",
     "NetworkTypedDict": ".create_router_responseop",
     "OrqAiTool": ".create_router_responseop",
@@ -17517,8 +17139,6 @@ _dynamic_imports: dict[str, str] = {
     "ToolsFilesTypedDict": ".create_router_responseop",
     "ToolsFunction": ".create_router_responseop",
     "ToolsFunctionTypedDict": ".create_router_responseop",
-    "ToolsMCPTool": ".create_router_responseop",
-    "ToolsMCPToolTypedDict": ".create_router_responseop",
     "ToolsMode": ".create_router_responseop",
     "ToolsTTL": ".create_router_responseop",
     "ToolsType": ".create_router_responseop",
@@ -18795,18 +18415,6 @@ _dynamic_imports: dict[str, str] = {
     "CreateManagementKeyRequestTypedDict": ".createmanagementkeyrequest",
     "CreateManagementKeyResponse": ".createmanagementkeyresponse",
     "CreateManagementKeyResponseTypedDict": ".createmanagementkeyresponse",
-    "CreateMcpGatewayRequest": ".createmcpgatewayrequest",
-    "CreateMcpGatewayRequestTypedDict": ".createmcpgatewayrequest",
-    "CreateMcpGatewayResponse": ".createmcpgatewayresponse",
-    "CreateMcpGatewayResponseTypedDict": ".createmcpgatewayresponse",
-    "CreateMcpServerRequest": ".createmcpserverrequest",
-    "CreateMcpServerRequestTypedDict": ".createmcpserverrequest",
-    "CreateMcpServerResponse": ".createmcpserverresponse",
-    "CreateMcpServerResponseTypedDict": ".createmcpserverresponse",
-    "CreateMcpToolsetRequest": ".createmcptoolsetrequest",
-    "CreateMcpToolsetRequestTypedDict": ".createmcptoolsetrequest",
-    "CreateMcpToolsetResponse": ".createmcptoolsetresponse",
-    "CreateMcpToolsetResponseTypedDict": ".createmcptoolsetresponse",
     "CreateMemoryDocumentRequest": ".creatememorydocumentop",
     "CreateMemoryDocumentRequestBody": ".creatememorydocumentop",
     "CreateMemoryDocumentRequestBodyTypedDict": ".creatememorydocumentop",
@@ -19544,12 +19152,6 @@ _dynamic_imports: dict[str, str] = {
     "DeleteKnowledgeRequestTypedDict": ".deleteknowledgeop",
     "DeleteManagementKeyResponse": ".deletemanagementkeyresponse",
     "DeleteManagementKeyResponseTypedDict": ".deletemanagementkeyresponse",
-    "DeleteMcpGatewayResponse": ".deletemcpgatewayresponse",
-    "DeleteMcpGatewayResponseTypedDict": ".deletemcpgatewayresponse",
-    "DeleteMcpServerResponse": ".deletemcpserverresponse",
-    "DeleteMcpServerResponseTypedDict": ".deletemcpserverresponse",
-    "DeleteMcpToolsetResponse": ".deletemcptoolsetresponse",
-    "DeleteMcpToolsetResponseTypedDict": ".deletemcptoolsetresponse",
     "DeleteMemoryDocumentRequest": ".deletememorydocumentop",
     "DeleteMemoryDocumentRequestTypedDict": ".deletememorydocumentop",
     "DeleteMemoryRequest": ".deletememoryop",
@@ -20804,12 +20406,6 @@ _dynamic_imports: dict[str, str] = {
     "GetLogResponseTypedDict": ".getlogresponse",
     "GetManagementKeyResponse": ".getmanagementkeyresponse",
     "GetManagementKeyResponseTypedDict": ".getmanagementkeyresponse",
-    "GetMcpGatewayResponse": ".getmcpgatewayresponse",
-    "GetMcpGatewayResponseTypedDict": ".getmcpgatewayresponse",
-    "GetMcpServerResponse": ".getmcpserverresponse",
-    "GetMcpServerResponseTypedDict": ".getmcpserverresponse",
-    "GetMcpToolsetResponse": ".getmcptoolsetresponse",
-    "GetMcpToolsetResponseTypedDict": ".getmcptoolsetresponse",
     "GetMonitorResponse": ".getmonitorresponse",
     "GetMonitorResponseTypedDict": ".getmonitorresponse",
     "GetNotifierResponse": ".getnotifierresponse",
@@ -21977,18 +21573,6 @@ _dynamic_imports: dict[str, str] = {
     "ListManagementCapabilitiesResponseTypedDict": ".listmanagementcapabilitiesresponse",
     "ListManagementKeysResponse": ".listmanagementkeysresponse",
     "ListManagementKeysResponseTypedDict": ".listmanagementkeysresponse",
-    "ListMcpGatewayCallsResponse": ".listmcpgatewaycallsresponse",
-    "ListMcpGatewayCallsResponseTypedDict": ".listmcpgatewaycallsresponse",
-    "ListMcpGatewaysResponse": ".listmcpgatewaysresponse",
-    "ListMcpGatewaysResponseTypedDict": ".listmcpgatewaysresponse",
-    "ListMcpGatewayToolsResponse": ".listmcpgatewaytoolsresponse",
-    "ListMcpGatewayToolsResponseTypedDict": ".listmcpgatewaytoolsresponse",
-    "ListMcpServersResponse": ".listmcpserversresponse",
-    "ListMcpServersResponseTypedDict": ".listmcpserversresponse",
-    "ListMcpSessionsResponse": ".listmcpsessionsresponse",
-    "ListMcpSessionsResponseTypedDict": ".listmcpsessionsresponse",
-    "ListMcpToolsetsResponse": ".listmcptoolsetsresponse",
-    "ListMcpToolsetsResponseTypedDict": ".listmcptoolsetsresponse",
     "ListMonitorPresetsResponse": ".listmonitorpresetsresponse",
     "ListMonitorPresetsResponseTypedDict": ".listmonitorpresetsresponse",
     "ListMonitorsResponse": ".listmonitorsresponse",
@@ -22218,101 +21802,9 @@ _dynamic_imports: dict[str, str] = {
     "ManagementPermissionMode": ".managementpermissionmode",
     "McpAccess": ".mcpaccess",
     "McpAccessTypedDict": ".mcpaccess",
-    "McpAuthConfig": ".mcpauthconfig",
-    "McpAuthConfigTypedDict": ".mcpauthconfig",
-    "McpAuthType": ".mcpauthtype",
-    "McpConnection": ".mcpconnection",
-    "McpConnectionTypedDict": ".mcpconnection",
-    "McpConnectionType": ".mcpconnectiontype",
-    "McpEgressPolicy": ".mcpegresspolicy",
-    "McpEgressPolicyTypedDict": ".mcpegresspolicy",
-    "McpGateway": ".mcpgateway",
-    "McpGatewayTypedDict": ".mcpgateway",
-    "McpGatewayCall": ".mcpgatewaycall",
-    "McpGatewayCallTypedDict": ".mcpgatewaycall",
-    "McpGatewayCallListRequest": ".mcpgatewaycalllistop",
-    "McpGatewayCallListRequestTypedDict": ".mcpgatewaycalllistop",
-    "McpGatewayCallStatus": ".mcpgatewaycallstatus",
-    "McpGatewayDeleteRequest": ".mcpgatewaydeleteop",
-    "McpGatewayDeleteRequestTypedDict": ".mcpgatewaydeleteop",
-    "McpGatewayGetRequest": ".mcpgatewaygetop",
-    "McpGatewayGetRequestTypedDict": ".mcpgatewaygetop",
-    "McpGatewayListRequest": ".mcpgatewaylistop",
-    "McpGatewayListRequestTypedDict": ".mcpgatewaylistop",
-    "McpGatewayListToolsRequest": ".mcpgatewaylisttoolsop",
-    "McpGatewayListToolsRequestTypedDict": ".mcpgatewaylisttoolsop",
-    "McpGatewayMode": ".mcpgatewaymode",
-    "McpGatewayServerLink": ".mcpgatewayserverlink",
-    "McpGatewayServerLinkTypedDict": ".mcpgatewayserverlink",
-    "McpGatewayStatus": ".mcpgatewaystatus",
-    "InputSchema": ".mcpgatewaytool",
-    "InputSchemaTypedDict": ".mcpgatewaytool",
-    "McpGatewayTool": ".mcpgatewaytool",
-    "McpGatewayToolTypedDict": ".mcpgatewaytool",
-    "McpGatewayToolTestResult": ".mcpgatewaytooltestresult",
-    "McpGatewayToolTestResultTypedDict": ".mcpgatewaytooltestresult",
-    "Result": ".mcpgatewaytooltestresult",
-    "ResultTypedDict": ".mcpgatewaytooltestresult",
-    "McpGatewayUpdateRequest": ".mcpgatewayupdateop",
-    "McpGatewayUpdateRequestTypedDict": ".mcpgatewayupdateop",
-    "McpHeaderSecret": ".mcpheadersecret",
-    "McpHeaderSecretTypedDict": ".mcpheadersecret",
-    "McpOAuthConfig": ".mcpoauthconfig",
-    "McpOAuthConfigTypedDict": ".mcpoauthconfig",
-    "McpRuntimeLimits": ".mcpruntimelimits",
-    "McpRuntimeLimitsTypedDict": ".mcpruntimelimits",
-    "McpServer": ".mcpserver",
-    "McpServerTypedDict": ".mcpserver",
-    "McpServerDeleteRequest": ".mcpserverdeleteop",
-    "McpServerDeleteRequestTypedDict": ".mcpserverdeleteop",
-    "McpServerGetRequest": ".mcpservergetop",
-    "McpServerGetRequestTypedDict": ".mcpservergetop",
-    "McpServerListRequest": ".mcpserverlistop",
-    "McpServerListRequestTypedDict": ".mcpserverlistop",
-    "McpServerSharing": ".mcpserversharing",
-    "McpServerSharingTypedDict": ".mcpserversharing",
-    "McpServerSharingScope": ".mcpserversharingscope",
-    "McpServerStatus": ".mcpserverstatus",
-    "McpServerSyncRequest": ".mcpserversyncop",
-    "McpServerSyncRequestTypedDict": ".mcpserversyncop",
-    "McpServerTestToolRequest": ".mcpservertesttoolop",
-    "McpServerTestToolRequestTypedDict": ".mcpservertesttoolop",
-    "McpServerUpdateRequest": ".mcpserverupdateop",
-    "McpServerUpdateRequestTypedDict": ".mcpserverupdateop",
-    "McpSession": ".mcpsession",
-    "McpSessionTypedDict": ".mcpsession",
-    "McpSessionListRequest": ".mcpsessionlistop",
-    "McpSessionListRequestTypedDict": ".mcpsessionlistop",
-    "McpSessionRevokeRequest": ".mcpsessionrevokeop",
-    "McpSessionRevokeRequestTypedDict": ".mcpsessionrevokeop",
-    "McpSessionStatus": ".mcpsessionstatus",
-    "McpSyncState": ".mcpsyncstate",
-    "McpSyncStateTypedDict": ".mcpsyncstate",
-    "McpTestResult": ".mcptestresult",
-    "McpTestResultTypedDict": ".mcptestresult",
-    "McpTool": ".mcptool",
-    "McpToolInputSchema": ".mcptool",
-    "McpToolInputSchemaTypedDict": ".mcptool",
-    "McpToolTypedDict": ".mcptool",
-    "McpToolExposure": ".mcptoolexposure",
-    "McpToolExposureTypedDict": ".mcptoolexposure",
-    "McpToolExposureMode": ".mcptoolexposuremode",
     "McpToolInput": ".mcptoolinput",
     "McpToolInputType": ".mcptoolinput",
     "McpToolInputTypedDict": ".mcptoolinput",
-    "McpToolNaming": ".mcptoolnaming",
-    "McpToolset": ".mcptoolset",
-    "McpToolsetTypedDict": ".mcptoolset",
-    "McpToolsetDeleteRequest": ".mcptoolsetdeleteop",
-    "McpToolsetDeleteRequestTypedDict": ".mcptoolsetdeleteop",
-    "McpToolsetGetRequest": ".mcptoolsetgetop",
-    "McpToolsetGetRequestTypedDict": ".mcptoolsetgetop",
-    "McpToolsetListRequest": ".mcptoolsetlistop",
-    "McpToolsetListRequestTypedDict": ".mcptoolsetlistop",
-    "McpToolsetToolRef": ".mcptoolsettoolref",
-    "McpToolsetToolRefTypedDict": ".mcptoolsettoolref",
-    "McpToolsetUpdateRequest": ".mcptoolsetupdateop",
-    "McpToolsetUpdateRequestTypedDict": ".mcptoolsetupdateop",
     "MemoryParam": ".memoryparam",
     "MemoryParamTypedDict": ".memoryparam",
     "ModelAzureFoundryDeploymentsRequestBody": ".modelazurefoundrydeploymentsop",
@@ -23355,8 +22847,6 @@ _dynamic_imports: dict[str, str] = {
     "RetrieveToolResponseBodyType": ".retrievetoolop",
     "RetrieveToolResponseBodyTypedDict": ".retrievetoolop",
     "ReviewOutcome": ".reviewoutcome",
-    "RevokeMcpSessionResponse": ".revokemcpsessionresponse",
-    "RevokeMcpSessionResponseTypedDict": ".revokemcpsessionresponse",
     "RoutingRule": ".routingrule",
     "RoutingRuleTypedDict": ".routingrule",
     "RoutingRuleCreateRequestBody": ".routingrulecreateop",
@@ -23424,6 +22914,8 @@ _dynamic_imports: dict[str, str] = {
     "AgentToolInputRunWebScraperToolTypedDict": ".runagentop",
     "AgentToolInputRunWriteMemoryStoreTool": ".runagentop",
     "AgentToolInputRunWriteMemoryStoreToolTypedDict": ".runagentop",
+    "Arguments": ".runagentop",
+    "ArgumentsTypedDict": ".runagentop",
     "Blueprint": ".runagentop",
     "BlueprintTypedDict": ".runagentop",
     "CodeTool": ".runagentop",
@@ -23472,8 +22964,6 @@ _dynamic_imports: dict[str, str] = {
     "RunAgentAgentToolInputRunAgentsRequestRequestBodyType": ".runagentop",
     "RunAgentAgentToolInputRunAgentsRequestType": ".runagentop",
     "RunAgentAgentToolInputRunAgentsType": ".runagentop",
-    "RunAgentAgentToolInputRunArguments": ".runagentop",
-    "RunAgentAgentToolInputRunArgumentsTypedDict": ".runagentop",
     "RunAgentAgentToolInputRunHeaders": ".runagentop",
     "RunAgentAgentToolInputRunHeadersTypedDict": ".runagentop",
     "RunAgentAgentToolInputRunParameters": ".runagentop",
@@ -24089,25 +23579,10 @@ _dynamic_imports: dict[str, str] = {
     "StreamRunAgentToolChoiceFunction": ".streamrunagentop",
     "StreamRunAgentToolChoiceFunctionTypedDict": ".streamrunagentop",
     "StreamRunAgentToolChoiceType": ".streamrunagentop",
-    "SyncMcpServerRequest": ".syncmcpserverrequest",
-    "SyncMcpServerRequestTypedDict": ".syncmcpserverrequest",
-    "SyncMcpServerResponse": ".syncmcpserverresponse",
-    "SyncMcpServerResponseTypedDict": ".syncmcpserverresponse",
-    "SyncStatus": ".syncstatus",
     "Telemetry": ".telemetry",
     "TelemetryTypedDict": ".telemetry",
     "TelemetryRow": ".telemetryrow",
     "TelemetryRowTypedDict": ".telemetryrow",
-    "TestMcpServerRequest": ".testmcpserverrequest",
-    "TestMcpServerRequestTypedDict": ".testmcpserverrequest",
-    "TestMcpServerResponse": ".testmcpserverresponse",
-    "TestMcpServerResponseTypedDict": ".testmcpserverresponse",
-    "Arguments": ".testmcpservertoolrequest",
-    "ArgumentsTypedDict": ".testmcpservertoolrequest",
-    "TestMcpServerToolRequest": ".testmcpservertoolrequest",
-    "TestMcpServerToolRequestTypedDict": ".testmcpservertoolrequest",
-    "TestMcpServerToolResponse": ".testmcpservertoolresponse",
-    "TestMcpServerToolResponseTypedDict": ".testmcpservertoolresponse",
     "CacheControl": ".textcontentpartschema",
     "CacheControlTypedDict": ".textcontentpartschema",
     "TTL": ".textcontentpartschema",
@@ -24825,18 +24300,6 @@ _dynamic_imports: dict[str, str] = {
     "UpdateManagementKeyRequestTypedDict": ".updatemanagementkeyrequest",
     "UpdateManagementKeyResponse": ".updatemanagementkeyresponse",
     "UpdateManagementKeyResponseTypedDict": ".updatemanagementkeyresponse",
-    "UpdateMcpGatewayRequest": ".updatemcpgatewayrequest",
-    "UpdateMcpGatewayRequestTypedDict": ".updatemcpgatewayrequest",
-    "UpdateMcpGatewayResponse": ".updatemcpgatewayresponse",
-    "UpdateMcpGatewayResponseTypedDict": ".updatemcpgatewayresponse",
-    "UpdateMcpServerRequest": ".updatemcpserverrequest",
-    "UpdateMcpServerRequestTypedDict": ".updatemcpserverrequest",
-    "UpdateMcpServerResponse": ".updatemcpserverresponse",
-    "UpdateMcpServerResponseTypedDict": ".updatemcpserverresponse",
-    "UpdateMcpToolsetRequest": ".updatemcptoolsetrequest",
-    "UpdateMcpToolsetRequestTypedDict": ".updatemcptoolsetrequest",
-    "UpdateMcpToolsetResponse": ".updatemcptoolsetresponse",
-    "UpdateMcpToolsetResponseTypedDict": ".updatemcptoolsetresponse",
     "UpdateMemoryDocumentRequest": ".updatememorydocumentop",
     "UpdateMemoryDocumentRequestBody": ".updatememorydocumentop",
     "UpdateMemoryDocumentRequestBodyTypedDict": ".updatememorydocumentop",