+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-30T10:46:39.932Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-30T11:26:10.323Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")