+++ packages/orq-rc/RELEASES.md
@@ -9869,3 +9869,13 @@ Based on:
 - [python v4.13.0-rc.20] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.20 - packages/orq-rc
+
+## 2026-07-30 11:29:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.21] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.21 - packages/orq-rc